coupon id added

+++ Promotions/views/v1/views.py
@@ -6590,10 +6590,11 @@ def create_coupon(request):
 
 @api_view(['DELETE'])
 @permission_classes([AllowAny])
-def delete_coupon(request):
-    coupon = Coupon.objects.all()
-    coupon.delete()
-    return Response({"msg": "Coupon deleted successfully"}, status=status.HTTP_200_OK)
+def delete_coupon(request,id=None):
+    if id:
+        coupon = Coupon.objects.filter(id=id)
+        coupon.delete()
+        return Response({"msg": "Coupon deleted successfully"}, status=status.HTTP_200_OK)
 
 
 @api_view(['GET'])
